+++ examples/data-sources/meraki_appliance_warm_spare/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_appliance_warm_spare" "example" {
+  network_id = "L_123456"
+}