+++ examples/resources/meraki_appliance_warm_spare/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_warm_spare.example "<network_id>"