+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - Add `meraki_appliance_site_to_site_vpn` resource and data source
 - Add `meraki_appliance_third_party_vpn_peers` resource and data source
 - Add `meraki_appliance_vpn_firewall_rules` resource and data source
+- Add `meraki_appliance_warm_spare` resource and data source
 
 ## 0.1.1